+++ pypeit/traceslits.py
@@ -1019,8 +1019,8 @@ class TraceSlits(masterframe.MasterFrame):
           Or have width less than fracignore
 
         Args:
+            plate_scale: float
             trim_slits: bool, optional
-            plate_scale: float, optional
             ech_slit_tol: float, optional
 
         Returns:
@@ -1439,8 +1439,7 @@ def load_traceslit_files(root):
     # Open FITS
     fits_file = root+'.fits.gz'
     if not os.path.isfile(fits_file):
-        msgs.warn("No TraceSlits FITS file found.  Returning None, None")
-        return None, None
+        msgs.error("No TraceSlits FITS file found!")
 
     msgs.info("Loading a pre-existing master calibration frame of type: trace from filename: {:}".format(fits_file))
     hdul = fits.open(fits_file)